--- src/gdpr/modal.ts
@@ -1,6 +1,0 @@
-import { createModal } from "../../Modal";
-
-export const modal = createModal({
-    "isOpenedByDefault": false,
-    "id": "fr-consent-modal"
-});
